cleanup(kernel_crawler): refactored common git related code between minikube and bottlerocket.
